sound: opl3: Add a KUnit test suite for the opl3
Add a very basic test suite for the OPL3's initialisation process.

There are three tests, each of which test part of the init and reset
process:
- opl3_test_init
- opl3_test_hwdep_new
- opl3_test_reset

These tests currently print out a dump of all register writes, and
verify some of the more important ones.

This patch is still a work-in-progress.

Signed-off-by: David Gow <david@ingeniumdigital.com>