sound: opl3: Add a driver for the OPL{2,3}LPT
The OPL2LPT and OPL3LPT are parallel port soundcards using the OPL2 and
OPL3 chips respecively.

This driver (when manually loaded for the appropriate parallel port)
permits these devices to be used for MIDI playback (if instruments are
loaded with sbiload), or via the ALSA hwdep FM interface.

See:
- https://www.serdashop.com/OPL2LPT
- https://www.serdashop.com/OPL3LPT
- https://davidgow.net/hacks/oplhw.html

Signed-off-by: David Gow <david@ingeniumdigital.com>