arch: x86: Implement mod/div compiler builtins
While the kernel deliberately doesn't implement compiler builtins for
64-bit division and modulus on 32-bit x86, it seems that rust may
require them for some standard library code.

Add basic implementations which are based on the math64.h macros.

Note that these currently live in moddiv_32.c, because they're for
32-bit x86, even though they operate on 64-bit integers (confusing,
isn't it). The right solution is probably to put them in rust's
compiler_builtins implentation, with all of its symbol hiding, etc.

Signed-off-by: David Gow <davidgow@google.com>