rust: kunit: Add test/test_suite macros, sample test
This is an in-progress patch to allow (and demonstrate) the creation of
100% rust KUnit tests and test modules.

Unlike the current doctests implementation, the various test case and
test suite structures (as well as the entires in a special section) are
generated from rust code, rather than a c helper file.

Note that this uses a bunch of unsafe, the const_mut_ref feature, and
the test suite macro doesn't fully work yet. (The whole thing is also a
bit outdated). This is largely due to me not knowing enough rust yet to
deal with the weirder unsafe / ffi bits.

Signed-off-by: David Gow <davidgow@google.com>